+++ src/agents/sessions/session-manager-core.ts
@@ -1,8 +1,8 @@
 import {
   loadTranscriptEventsSync,
   replaceTranscriptEventsSync,
+  type SessionTranscriptRuntimeTarget,
 } from "../../config/sessions/session-accessor.js";
-import type { SessionTranscriptRuntimeTarget } from "../../config/sessions/session-accessor.types.js";
 import { isSessionTranscriptSideAppendEntry } from "../../config/sessions/transcript-tree.js";
 import { CURRENT_SESSION_VERSION } from "../../config/sessions/version.js";
 import {
